+++ src/com/morgan/design/activity/OrmLiteFragment.java
@@ -7,21 +7,21 @@ import com.morgan.design.db.DatabaseHelper;
 
 public class OrmLiteFragment extends Fragment {
 
-	private DatabaseHelper databaseHelper = null;
+    private DatabaseHelper databaseHelper = null;
 
-	protected DatabaseHelper getHelper() {
-		if (databaseHelper == null) {
-			databaseHelper = OpenHelperManager.getHelper(getActivity(), DatabaseHelper.class);
-		}
-		return databaseHelper;
-	}
+    protected DatabaseHelper getHelper() {
+        if (databaseHelper == null) {
+            databaseHelper = OpenHelperManager.getHelper(getActivity(), DatabaseHelper.class);
+        }
+        return databaseHelper;
+    }
 
-	@Override
-	public void onDestroy() {
-		super.onDestroy();
-		if (databaseHelper != null) {
-			OpenHelperManager.releaseHelper();
-			databaseHelper = null;
-		}
-	}
+    @Override
+    public void onDestroy() {
+        super.onDestroy();
+        if (databaseHelper != null) {
+            OpenHelperManager.releaseHelper();
+            databaseHelper = null;
+        }
+    }
 }
