+++ src/com/morgan/design/activity/preferences/SettingsActivity.java
@@ -5,10 +5,10 @@ import android.os.Bundle;
 
 public class SettingsActivity extends Activity {
 
-	@Override
-	protected void onCreate(Bundle savedInstanceState) {
-		super.onCreate(savedInstanceState);
-		getFragmentManager().beginTransaction().replace(android.R.id.content, new PreferencesFragment()).commit();
-	}
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        getFragmentManager().beginTransaction().replace(android.R.id.content, new PreferencesFragment()).commit();
+    }
 
 }
