+++ src/com/morgan/design/helpers/Changelog.java
@@ -1,8 +1,5 @@
 package com.morgan.design.helpers;
 
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
 import android.app.Activity;
 import android.app.AlertDialog;
 import android.content.DialogInterface;
@@ -14,49 +11,51 @@ import com.morgan.design.utils.BuildUtils;
 import com.morgan.design.utils.Prefs;
 import com.morgan.design.utils.Utils;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
 public class Changelog {
 
-	private final static Logger LOG = LoggerFactory.getLogger(Changelog.class);
+    private final static Logger LOG = LoggerFactory.getLogger(Changelog.class);
 
-	public static boolean show(final Activity activity) {
+    public static boolean show(final Activity activity) {
 
-		final int prefVersion = Prefs.getAppVersionPref(activity);
-		int currentVersion = 0;
+        final int prefVersion = Prefs.getAppVersionPref(activity);
+        int currentVersion = 0;
 
-		final boolean overrideChangeLog = BuildUtils.isRunningEmmulator();
-		LOG.debug("Overriding ChangeLog: {}", overrideChangeLog);
+        final boolean overrideChangeLog = BuildUtils.isRunningEmmulator();
+        LOG.debug("Overriding ChangeLog: {}", overrideChangeLog);
 
-		final boolean showChangeLog = Prefs.getChangelogPref(activity);
+        final boolean showChangeLog = Prefs.getChangelogPref(activity);
 
-		try {
-			final PackageInfo pi = activity.getPackageManager().getPackageInfo(activity.getPackageName(), 0);
-			currentVersion = pi.versionCode;
-		}
-		catch (final NameNotFoundException e) {
-			LOG.error("Package name not found", e);
-			return false;
-		}
+        try {
+            final PackageInfo pi = activity.getPackageManager().getPackageInfo(activity.getPackageName(), 0);
+            currentVersion = pi.versionCode;
+        } catch (final NameNotFoundException e) {
+            LOG.error("Package name not found", e);
+            return false;
+        }
 
-		// Not added before
-		if (prefVersion == 0 && showChangeLog) {
-			showChangelogDialog(activity);
-		}
-		// On upgrade
-		else if (overrideChangeLog || (showChangeLog && currentVersion > prefVersion)) {
-			showChangelogDialog(activity);
-		}
-		Prefs.setAppVersionPref(activity, currentVersion);
-		return true;
-	}
+        // Not added before
+        if (prefVersion == 0 && showChangeLog) {
+            showChangelogDialog(activity);
+        }
+        // On upgrade
+        else if (overrideChangeLog || (showChangeLog && currentVersion > prefVersion)) {
+            showChangelogDialog(activity);
+        }
+        Prefs.setAppVersionPref(activity, currentVersion);
+        return true;
+    }
 
-	protected static void showChangelogDialog(final Activity activity) {
-		new AlertDialog.Builder(activity).setIcon(android.R.drawable.ic_dialog_info).setTitle(R.string.changelog_title)
-				.setView(Utils.dialogWebView(activity, activity.getString(R.string.changelog_filename)))
-				.setPositiveButton(R.string.ok, null).setNegativeButton(R.string.feedback, new DialogInterface.OnClickListener() {
-					@Override
-					public void onClick(final DialogInterface dialog, final int which) {
-						Utils.openFeedback(activity);
-					}
-				}).show();
-	}
+    protected static void showChangelogDialog(final Activity activity) {
+        new AlertDialog.Builder(activity).setIcon(android.R.drawable.ic_dialog_info).setTitle(R.string.changelog_title)
+                .setView(Utils.dialogWebView(activity, activity.getString(R.string.changelog_filename)))
+                .setPositiveButton(R.string.ok, null).setNegativeButton(R.string.feedback, new DialogInterface.OnClickListener() {
+            @Override
+            public void onClick(final DialogInterface dialog, final int which) {
+                Utils.openFeedback(activity);
+            }
+        }).show();
+    }
 }
