+++ src/com/morgan/design/helpers/Constants.java
@@ -2,15 +2,15 @@ package com.morgan.design.helpers;
 
 public class Constants {
 
-	public static final String[] MY_EMAIL = new String[] { "james.morgan.e@googlemail.com" };
-	public static final String[] APP_FEEDBACK_EMAIL = new String[] { "appfeedback@morgan-design.com" };
+    public static final String[] MY_EMAIL = new String[]{"james.morgan.e@googlemail.com"};
+    public static final String[] APP_FEEDBACK_EMAIL = new String[]{"appfeedback@morgan-design.com"};
 
-	public static final String DONATE_URL = "https://www.paypal.com/cgi-bin/webscr?cmd=_s-xclick&hosted_button_id=QMZWPL26PY5GW";
-	public static final String MORGAN_DESIGN = "http://www.morgan-design.com";
-	public static final String TWITTER_URL = "http://www.twitter.com/jimbob_87";
-	public static final String MARKET_LINK_URL = "market://search?q=MorganDesign&c=apps";
-	public static final String TANGO_ICON_URL = "http://tango.freedesktop.org/Tango_Icon_Library";
+    public static final String DONATE_URL = "https://www.paypal.com/cgi-bin/webscr?cmd=_s-xclick&hosted_button_id=QMZWPL26PY5GW";
+    public static final String MORGAN_DESIGN = "http://www.morgan-design.com";
+    public static final String TWITTER_URL = "http://www.twitter.com/jimbob_87";
+    public static final String MARKET_LINK_URL = "market://search?q=MorganDesign&c=apps";
+    public static final String TANGO_ICON_URL = "http://tango.freedesktop.org/Tango_Icon_Library";
 
-	public static final String GOOGLE_ANALYTIC_ACCOUNT_ID = "UA-22912453-1";
+    public static final String GOOGLE_ANALYTIC_ACCOUNT_ID = "UA-22912453-1";
 
 }
