+++ src/com/morgan/design/helpers/Credit.java
@@ -2,41 +2,41 @@ package com.morgan.design.helpers;
 
 public class Credit {
 
-	private String title;
-	private String caption;
-	private String creditWebAddress;
-	private int logo;
+    private String title;
+    private String caption;
+    private String creditWebAddress;
+    private int logo;
 
-	public String getTitle() {
-		return title;
-	}
+    public String getTitle() {
+        return title;
+    }
 
-	public void setTitle(String title) {
-		this.title = title;
-	}
+    public void setTitle(String title) {
+        this.title = title;
+    }
 
-	public String getCaption() {
-		return caption;
-	}
+    public String getCaption() {
+        return caption;
+    }
 
-	public void setCaption(String caption) {
-		this.caption = caption;
-	}
+    public void setCaption(String caption) {
+        this.caption = caption;
+    }
 
-	public String getCreditWebAddress() {
-		return creditWebAddress;
-	}
+    public String getCreditWebAddress() {
+        return creditWebAddress;
+    }
 
-	public void setCreditWebAddress(String creditWebAddress) {
-		this.creditWebAddress = creditWebAddress;
-	}
+    public void setCreditWebAddress(String creditWebAddress) {
+        this.creditWebAddress = creditWebAddress;
+    }
 
-	public int getLogo() {
-		return logo;
-	}
+    public int getLogo() {
+        return logo;
+    }
 
-	public void setLogo(int logo) {
-		this.logo = logo;
-	}
+    public void setLogo(int logo) {
+        this.logo = logo;
+    }
 
 }
