+++ src/com/morgan/design/ui/ObservableAdapter.java
@@ -1,5 +1,5 @@
 package com.morgan.design.ui;
 
 public interface ObservableAdapter {
-	public void removeObserver();
+    public void removeObserver();
 }
