+++ src/com/morgan/design/utils/InputUtils.java
@@ -6,14 +6,14 @@ import android.view.inputmethod.InputMethodManager;
 
 public class InputUtils {
 
-	public static void hideKeyBoard(final View view, final Context context) {
-		final InputMethodManager imm = (InputMethodManager) context.getSystemService(Context.INPUT_METHOD_SERVICE);
-		imm.hideSoftInputFromWindow(view.getWindowToken(), 0);
-	}
+    public static void hideKeyBoard(final View view, final Context context) {
+        final InputMethodManager imm = (InputMethodManager) context.getSystemService(Context.INPUT_METHOD_SERVICE);
+        imm.hideSoftInputFromWindow(view.getWindowToken(), 0);
+    }
 
-	public static void showKeyBoard(final View view, final Context context) {
-		final InputMethodManager imm = (InputMethodManager) context.getSystemService(Context.INPUT_METHOD_SERVICE);
-		imm.showSoftInput(view, InputMethodManager.SHOW_FORCED);
-	}
+    public static void showKeyBoard(final View view, final Context context) {
+        final InputMethodManager imm = (InputMethodManager) context.getSystemService(Context.INPUT_METHOD_SERVICE);
+        imm.showSoftInput(view, InputMethodManager.SHOW_FORCED);
+    }
 
 }
