+++ src/com/morgan/design/utils/StringUtils.java
@@ -13,7 +13,7 @@ package com.morgan.design.utils;
  * <p>
  * Operations on {@link java.lang.String} that are <code>null</code> safe.
  * </p>
- * 
+ * <p/>
  * <ul>
  * <li><b>IsEmpty/IsBlank</b> - checks if a String contains text</li>
  * <li><b>Trim/Strip</b> - removes leading and trailing whitespace</li>
@@ -39,11 +39,11 @@ package com.morgan.design.utils;
  * <li><b>Difference</b> - compares Strings and reports on their differences</li>
  * <li><b>LevensteinDistance</b> - the number of changes needed to change one String into another</li>
  * </ul>
- * 
+ * <p/>
  * <p>
  * The <code>StringUtils</code> class defines certain words related to String handling.
  * </p>
- * 
+ * <p/>
  * <ul>
  * <li>null - <code>null</code></li>
  * <li>empty - a zero-length string (<code>""</code>)</li>
@@ -51,28 +51,27 @@ package com.morgan.design.utils;
  * <li>whitespace - the characters defined by {@link Character#isWhitespace(char)}</li>
  * <li>trim - the characters &lt;= 32 as in {@link String#trim()}</li>
  * </ul>
- * 
+ * <p/>
  * <p>
  * <code>StringUtils</code> handles <code>null</code> input Strings quietly. That is to say that a <code>null</code>
  * input will return <code>null</code>. Where a <code>boolean</code> or <code>int</code> is being returned details vary
  * by method.
  * </p>
- * 
+ * <p/>
  * <p>
  * A side effect of the <code>null</code> handling is that a <code>NullPointerException</code> should be considered a
  * bug in <code>StringUtils</code> (except for deprecated methods).
  * </p>
- * 
+ * <p/>
  * <p>
  * Methods in this class give sample code to explain their operation. The symbol <code>*</code> is used to indicate any
  * input including <code>null</code>.
  * </p>
- * 
+ * <p/>
  * <p>
  * #ThreadSafe#
  * </p>
- * 
- * @see java.lang.String
+ *
  * @author Apache Software Foundation
  * @author <a href="http://jakarta.apache.org/turbine/">Apache Jakarta Turbine</a>
  * @author <a href="mailto:jon@latchkey.com">Jon S. Stevens</a>
@@ -92,558 +91,525 @@ package com.morgan.design.utils;
  * @author Reuben Sivan
  * @author Chris Hyzer
  * @author Scott Johnson
- * @since 1.0
  * @version $Id: StringUtils.java 1058365 2011-01-13 00:04:49Z niallp $
+ * @see java.lang.String
+ * @since 1.0
  */
 // @Immutable
 public class StringUtils {
-	// Performance testing notes (JDK 1.4, Jul03, scolebourne)
-	// Whitespace:
-	// Character.isWhitespace() is faster than WHITESPACE.indexOf()
-	// where WHITESPACE is a string of all whitespace characters
-	//
-	// Character access:
-	// String.charAt(n) versus toCharArray(), then array[n]
-	// String.charAt(n) is about 15% worse for a 10K string
-	// They are about equal for a length 50 string
-	// String.charAt(n) is about 4 times better for a length 3 string
-	// String.charAt(n) is best bet overall
-	//
-	// Append:
-	// String.concat about twice as fast as StringBuffer.append
-	// (not sure who tested this)
+    // Performance testing notes (JDK 1.4, Jul03, scolebourne)
+    // Whitespace:
+    // Character.isWhitespace() is faster than WHITESPACE.indexOf()
+    // where WHITESPACE is a string of all whitespace characters
+    //
+    // Character access:
+    // String.charAt(n) versus toCharArray(), then array[n]
+    // String.charAt(n) is about 15% worse for a 10K string
+    // They are about equal for a length 50 string
+    // String.charAt(n) is about 4 times better for a length 3 string
+    // String.charAt(n) is best bet overall
+    //
+    // Append:
+    // String.concat about twice as fast as StringBuffer.append
+    // (not sure who tested this)
 
-	/**
-	 * The empty String <code>""</code>.
-	 * 
-	 * @since 2.0
-	 */
-	public static final String EMPTY = "";
+    /**
+     * The empty String <code>""</code>.
+     *
+     * @since 2.0
+     */
+    public static final String EMPTY = "";
 
-	/**
-	 * Represents a failed index search.
-	 * 
-	 * @since 2.1
-	 */
-	public static final int INDEX_NOT_FOUND = -1;
+    /**
+     * Represents a failed index search.
+     *
+     * @since 2.1
+     */
+    public static final int INDEX_NOT_FOUND = -1;
 
-	/**
-	 * <p>
-	 * The maximum size to which the padding constant(s) can expand.
-	 * </p>
-	 */
-	private static final int PAD_LIMIT = 8192;
+    /**
+     * <p>
+     * The maximum size to which the padding constant(s) can expand.
+     * </p>
+     */
+    private static final int PAD_LIMIT = 8192;
 
-	/**
-	 * <p>
-	 * <code>StringUtils</code> instances should NOT be constructed in standard programming. Instead, the class should
-	 * be used as <code>StringUtils.trim(" foo ");</code>.
-	 * </p>
-	 * 
-	 * <p>
-	 * This constructor is public to permit tools that require a JavaBean instance to operate.
-	 * </p>
-	 */
-	public StringUtils() {
-		super();
-	}
+    /**
+     * <p>
+     * <code>StringUtils</code> instances should NOT be constructed in standard programming. Instead, the class should
+     * be used as <code>StringUtils.trim(" foo ");</code>.
+     * </p>
+     * <p/>
+     * <p>
+     * This constructor is public to permit tools that require a JavaBean instance to operate.
+     * </p>
+     */
+    public StringUtils() {
+        super();
+    }
 
-	// Empty checks
-	// -----------------------------------------------------------------------
-	/**
-	 * <p>
-	 * Checks if a String is empty ("") or null.
-	 * </p>
-	 * 
-	 * <pre>
-	 * StringUtils.isEmpty(null)      = true
-	 * StringUtils.isEmpty("")        = true
-	 * StringUtils.isEmpty(" ")       = false
-	 * StringUtils.isEmpty("bob")     = false
-	 * StringUtils.isEmpty("  bob  ") = false
-	 * </pre>
-	 * 
-	 * <p>
-	 * NOTE: This method changed in Lang version 2.0. It no longer trims the String. That functionality is available in
-	 * isBlank().
-	 * </p>
-	 * 
-	 * @param str
-	 *            the String to check, may be null
-	 * @return <code>true</code> if the String is empty or null
-	 */
-	public static boolean isEmpty(String str) {
-		return str == null || str.length() == 0;
-	}
+    // Empty checks
+    // -----------------------------------------------------------------------
 
-	/**
-	 * <p>
-	 * Checks if a String is not empty ("") and not null.
-	 * </p>
-	 * 
-	 * <pre>
-	 * StringUtils.isNotEmpty(null)      = false
-	 * StringUtils.isNotEmpty("")        = false
-	 * StringUtils.isNotEmpty(" ")       = true
-	 * StringUtils.isNotEmpty("bob")     = true
-	 * StringUtils.isNotEmpty("  bob  ") = true
-	 * </pre>
-	 * 
-	 * @param str
-	 *            the String to check, may be null
-	 * @return <code>true</code> if the String is not empty and not null
-	 */
-	public static boolean isNotEmpty(String str) {
-		return !StringUtils.isEmpty(str);
-	}
+    /**
+     * <p>
+     * Checks if a String is empty ("") or null.
+     * </p>
+     * <p/>
+     * <pre>
+     * StringUtils.isEmpty(null)      = true
+     * StringUtils.isEmpty("")        = true
+     * StringUtils.isEmpty(" ")       = false
+     * StringUtils.isEmpty("bob")     = false
+     * StringUtils.isEmpty("  bob  ") = false
+     * </pre>
+     * <p/>
+     * <p>
+     * NOTE: This method changed in Lang version 2.0. It no longer trims the String. That functionality is available in
+     * isBlank().
+     * </p>
+     *
+     * @param str the String to check, may be null
+     * @return <code>true</code> if the String is empty or null
+     */
+    public static boolean isEmpty(String str) {
+        return str == null || str.length() == 0;
+    }
 
-	/**
-	 * <p>
-	 * Checks if a String is whitespace, empty ("") or null.
-	 * </p>
-	 * 
-	 * <pre>
-	 * StringUtils.isBlank(null)      = true
-	 * StringUtils.isBlank("")        = true
-	 * StringUtils.isBlank(" ")       = true
-	 * StringUtils.isBlank("bob")     = false
-	 * StringUtils.isBlank("  bob  ") = false
-	 * </pre>
-	 * 
-	 * @param str
-	 *            the String to check, may be null
-	 * @return <code>true</code> if the String is null, empty or whitespace
-	 * @since 2.0
-	 */
-	public static boolean isBlank(String str) {
-		int strLen;
-		if (str == null || (strLen = str.length()) == 0) {
-			return true;
-		}
-		for (int i = 0; i < strLen; i++) {
-			if ((Character.isWhitespace(str.charAt(i)) == false)) {
-				return false;
-			}
-		}
-		return true;
-	}
+    /**
+     * <p>
+     * Checks if a String is not empty ("") and not null.
+     * </p>
+     * <p/>
+     * <pre>
+     * StringUtils.isNotEmpty(null)      = false
+     * StringUtils.isNotEmpty("")        = false
+     * StringUtils.isNotEmpty(" ")       = true
+     * StringUtils.isNotEmpty("bob")     = true
+     * StringUtils.isNotEmpty("  bob  ") = true
+     * </pre>
+     *
+     * @param str the String to check, may be null
+     * @return <code>true</code> if the String is not empty and not null
+     */
+    public static boolean isNotEmpty(String str) {
+        return !StringUtils.isEmpty(str);
+    }
 
-	/**
-	 * <p>
-	 * Checks if a String is not empty (""), not null and not whitespace only.
-	 * </p>
-	 * 
-	 * <pre>
-	 * StringUtils.isNotBlank(null)      = false
-	 * StringUtils.isNotBlank("")        = false
-	 * StringUtils.isNotBlank(" ")       = false
-	 * StringUtils.isNotBlank("bob")     = true
-	 * StringUtils.isNotBlank("  bob  ") = true
-	 * </pre>
-	 * 
-	 * @param str
-	 *            the String to check, may be null
-	 * @return <code>true</code> if the String is not empty and not null and not whitespace
-	 * @since 2.0
-	 */
-	public static boolean isNotBlank(String str) {
-		return !StringUtils.isBlank(str);
-	}
+    /**
+     * <p>
+     * Checks if a String is whitespace, empty ("") or null.
+     * </p>
+     * <p/>
+     * <pre>
+     * StringUtils.isBlank(null)      = true
+     * StringUtils.isBlank("")        = true
+     * StringUtils.isBlank(" ")       = true
+     * StringUtils.isBlank("bob")     = false
+     * StringUtils.isBlank("  bob  ") = false
+     * </pre>
+     *
+     * @param str the String to check, may be null
+     * @return <code>true</code> if the String is null, empty or whitespace
+     * @since 2.0
+     */
+    public static boolean isBlank(String str) {
+        int strLen;
+        if (str == null || (strLen = str.length()) == 0) {
+            return true;
+        }
+        for (int i = 0; i < strLen; i++) {
+            if ((Character.isWhitespace(str.charAt(i)) == false)) {
+                return false;
+            }
+        }
+        return true;
+    }
 
-	/**
-	 * <p>
-	 * Returns padding using the specified delimiter repeated to a given length.
-	 * </p>
-	 * 
-	 * <pre>
-	 * StringUtils.padding(0, 'e')  = ""
-	 * StringUtils.padding(3, 'e')  = "eee"
-	 * StringUtils.padding(-2, 'e') = IndexOutOfBoundsException
-	 * </pre>
-	 * 
-	 * <p>
-	 * Note: this method doesn't not support padding with <a
-	 * href="http://www.unicode.org/glossary/#supplementary_character">Unicode Supplementary Characters</a> as they
-	 * require a pair of <code>char</code>s to be represented. If you are needing to support full I18N of your
-	 * applications consider using {@link #repeat(String, int)} instead.
-	 * </p>
-	 * 
-	 * @param repeat
-	 *            number of times to repeat delim
-	 * @param padChar
-	 *            character to repeat
-	 * @return String with repeated character
-	 * @throws IndexOutOfBoundsException
-	 *             if <code>repeat &lt; 0</code>
-	 * @see #repeat(String, int)
-	 */
-	private static String padding(int repeat, char padChar) throws IndexOutOfBoundsException {
-		if (repeat < 0) {
-			throw new IndexOutOfBoundsException("Cannot pad a negative amount: " + repeat);
-		}
-		final char[] buf = new char[repeat];
-		for (int i = 0; i < buf.length; i++) {
-			buf[i] = padChar;
-		}
-		return new String(buf);
-	}
+    /**
+     * <p>
+     * Checks if a String is not empty (""), not null and not whitespace only.
+     * </p>
+     * <p/>
+     * <pre>
+     * StringUtils.isNotBlank(null)      = false
+     * StringUtils.isNotBlank("")        = false
+     * StringUtils.isNotBlank(" ")       = false
+     * StringUtils.isNotBlank("bob")     = true
+     * StringUtils.isNotBlank("  bob  ") = true
+     * </pre>
+     *
+     * @param str the String to check, may be null
+     * @return <code>true</code> if the String is not empty and not null and not whitespace
+     * @since 2.0
+     */
+    public static boolean isNotBlank(String str) {
+        return !StringUtils.isBlank(str);
+    }
 
-	/**
-	 * <p>
-	 * Right pad a String with spaces (' ').
-	 * </p>
-	 * 
-	 * <p>
-	 * The String is padded to the size of <code>size</code>.
-	 * </p>
-	 * 
-	 * <pre>
-	 * StringUtils.rightPad(null, *)   = null
-	 * StringUtils.rightPad("", 3)     = "   "
-	 * StringUtils.rightPad("bat", 3)  = "bat"
-	 * StringUtils.rightPad("bat", 5)  = "bat  "
-	 * StringUtils.rightPad("bat", 1)  = "bat"
-	 * StringUtils.rightPad("bat", -1) = "bat"
-	 * </pre>
-	 * 
-	 * @param str
-	 *            the String to pad out, may be null
-	 * @param size
-	 *            the size to pad to
-	 * @return right padded String or original String if no padding is necessary, <code>null</code> if null String input
-	 */
-	public static String rightPad(String str, int size) {
-		return rightPad(str, size, ' ');
-	}
+    /**
+     * <p>
+     * Returns padding using the specified delimiter repeated to a given length.
+     * </p>
+     * <p/>
+     * <pre>
+     * StringUtils.padding(0, 'e')  = ""
+     * StringUtils.padding(3, 'e')  = "eee"
+     * StringUtils.padding(-2, 'e') = IndexOutOfBoundsException
+     * </pre>
+     * <p/>
+     * <p>
+     * Note: this method doesn't not support padding with <a
+     * href="http://www.unicode.org/glossary/#supplementary_character">Unicode Supplementary Characters</a> as they
+     * require a pair of <code>char</code>s to be represented. If you are needing to support full I18N of your
+     * applications consider using {@link #repeat(String, int)} instead.
+     * </p>
+     *
+     * @param repeat  number of times to repeat delim
+     * @param padChar character to repeat
+     * @return String with repeated character
+     * @throws IndexOutOfBoundsException if <code>repeat &lt; 0</code>
+     * @see #repeat(String, int)
+     */
+    private static String padding(int repeat, char padChar) throws IndexOutOfBoundsException {
+        if (repeat < 0) {
+            throw new IndexOutOfBoundsException("Cannot pad a negative amount: " + repeat);
+        }
+        final char[] buf = new char[repeat];
+        for (int i = 0; i < buf.length; i++) {
+            buf[i] = padChar;
+        }
+        return new String(buf);
+    }
 
-	/**
-	 * <p>
-	 * Right pad a String with a specified character.
-	 * </p>
-	 * 
-	 * <p>
-	 * The String is padded to the size of <code>size</code>.
-	 * </p>
-	 * 
-	 * <pre>
-	 * StringUtils.rightPad(null, *, *)     = null
-	 * StringUtils.rightPad("", 3, 'z')     = "zzz"
-	 * StringUtils.rightPad("bat", 3, 'z')  = "bat"
-	 * StringUtils.rightPad("bat", 5, 'z')  = "batzz"
-	 * StringUtils.rightPad("bat", 1, 'z')  = "bat"
-	 * StringUtils.rightPad("bat", -1, 'z') = "bat"
-	 * </pre>
-	 * 
-	 * @param str
-	 *            the String to pad out, may be null
-	 * @param size
-	 *            the size to pad to
-	 * @param padChar
-	 *            the character to pad with
-	 * @return right padded String or original String if no padding is necessary, <code>null</code> if null String input
-	 * @since 2.0
-	 */
-	public static String rightPad(String str, int size, char padChar) {
-		if (str == null) {
-			return null;
-		}
-		int pads = size - str.length();
-		if (pads <= 0) {
-			return str; // returns original String when possible
-		}
-		if (pads > PAD_LIMIT) {
-			return rightPad(str, size, String.valueOf(padChar));
-		}
-		return str.concat(padding(pads, padChar));
-	}
+    /**
+     * <p>
+     * Right pad a String with spaces (' ').
+     * </p>
+     * <p/>
+     * <p>
+     * The String is padded to the size of <code>size</code>.
+     * </p>
+     * <p/>
+     * <pre>
+     * StringUtils.rightPad(null, *)   = null
+     * StringUtils.rightPad("", 3)     = "   "
+     * StringUtils.rightPad("bat", 3)  = "bat"
+     * StringUtils.rightPad("bat", 5)  = "bat  "
+     * StringUtils.rightPad("bat", 1)  = "bat"
+     * StringUtils.rightPad("bat", -1) = "bat"
+     * </pre>
+     *
+     * @param str  the String to pad out, may be null
+     * @param size the size to pad to
+     * @return right padded String or original String if no padding is necessary, <code>null</code> if null String input
+     */
+    public static String rightPad(String str, int size) {
+        return rightPad(str, size, ' ');
+    }
 
-	/**
-	 * <p>
-	 * Right pad a String with a specified String.
-	 * </p>
-	 * 
-	 * <p>
-	 * The String is padded to the size of <code>size</code>.
-	 * </p>
-	 * 
-	 * <pre>
-	 * StringUtils.rightPad(null, *, *)      = null
-	 * StringUtils.rightPad("", 3, "z")      = "zzz"
-	 * StringUtils.rightPad("bat", 3, "yz")  = "bat"
-	 * StringUtils.rightPad("bat", 5, "yz")  = "batyz"
-	 * StringUtils.rightPad("bat", 8, "yz")  = "batyzyzy"
-	 * StringUtils.rightPad("bat", 1, "yz")  = "bat"
-	 * StringUtils.rightPad("bat", -1, "yz") = "bat"
-	 * StringUtils.rightPad("bat", 5, null)  = "bat  "
-	 * StringUtils.rightPad("bat", 5, "")    = "bat  "
-	 * </pre>
-	 * 
-	 * @param str
-	 *            the String to pad out, may be null
-	 * @param size
-	 *            the size to pad to
-	 * @param padStr
-	 *            the String to pad with, null or empty treated as single space
-	 * @return right padded String or original String if no padding is necessary, <code>null</code> if null String input
-	 */
-	public static String rightPad(String str, int size, String padStr) {
-		if (str == null) {
-			return null;
-		}
-		if (isEmpty(padStr)) {
-			padStr = " ";
-		}
-		int padLen = padStr.length();
-		int strLen = str.length();
-		int pads = size - strLen;
-		if (pads <= 0) {
-			return str; // returns original String when possible
-		}
-		if (padLen == 1 && pads <= PAD_LIMIT) {
-			return rightPad(str, size, padStr.charAt(0));
-		}
+    /**
+     * <p>
+     * Right pad a String with a specified character.
+     * </p>
+     * <p/>
+     * <p>
+     * The String is padded to the size of <code>size</code>.
+     * </p>
+     * <p/>
+     * <pre>
+     * StringUtils.rightPad(null, *, *)     = null
+     * StringUtils.rightPad("", 3, 'z')     = "zzz"
+     * StringUtils.rightPad("bat", 3, 'z')  = "bat"
+     * StringUtils.rightPad("bat", 5, 'z')  = "batzz"
+     * StringUtils.rightPad("bat", 1, 'z')  = "bat"
+     * StringUtils.rightPad("bat", -1, 'z') = "bat"
+     * </pre>
+     *
+     * @param str     the String to pad out, may be null
+     * @param size    the size to pad to
+     * @param padChar the character to pad with
+     * @return right padded String or original String if no padding is necessary, <code>null</code> if null String input
+     * @since 2.0
+     */
+    public static String rightPad(String str, int size, char padChar) {
+        if (str == null) {
+            return null;
+        }
+        int pads = size - str.length();
+        if (pads <= 0) {
+            return str; // returns original String when possible
+        }
+        if (pads > PAD_LIMIT) {
+            return rightPad(str, size, String.valueOf(padChar));
+        }
+        return str.concat(padding(pads, padChar));
+    }
 
-		if (pads == padLen) {
-			return str.concat(padStr);
-		}
-		else if (pads < padLen) {
-			return str.concat(padStr.substring(0, pads));
-		}
-		else {
-			char[] padding = new char[pads];
-			char[] padChars = padStr.toCharArray();
-			for (int i = 0; i < pads; i++) {
-				padding[i] = padChars[i % padLen];
-			}
-			return str.concat(new String(padding));
-		}
-	}
+    /**
+     * <p>
+     * Right pad a String with a specified String.
+     * </p>
+     * <p/>
+     * <p>
+     * The String is padded to the size of <code>size</code>.
+     * </p>
+     * <p/>
+     * <pre>
+     * StringUtils.rightPad(null, *, *)      = null
+     * StringUtils.rightPad("", 3, "z")      = "zzz"
+     * StringUtils.rightPad("bat", 3, "yz")  = "bat"
+     * StringUtils.rightPad("bat", 5, "yz")  = "batyz"
+     * StringUtils.rightPad("bat", 8, "yz")  = "batyzyzy"
+     * StringUtils.rightPad("bat", 1, "yz")  = "bat"
+     * StringUtils.rightPad("bat", -1, "yz") = "bat"
+     * StringUtils.rightPad("bat", 5, null)  = "bat  "
+     * StringUtils.rightPad("bat", 5, "")    = "bat  "
+     * </pre>
+     *
+     * @param str    the String to pad out, may be null
+     * @param size   the size to pad to
+     * @param padStr the String to pad with, null or empty treated as single space
+     * @return right padded String or original String if no padding is necessary, <code>null</code> if null String input
+     */
+    public static String rightPad(String str, int size, String padStr) {
+        if (str == null) {
+            return null;
+        }
+        if (isEmpty(padStr)) {
+            padStr = " ";
+        }
+        int padLen = padStr.length();
+        int strLen = str.length();
+        int pads = size - strLen;
+        if (pads <= 0) {
+            return str; // returns original String when possible
+        }
+        if (padLen == 1 && pads <= PAD_LIMIT) {
+            return rightPad(str, size, padStr.charAt(0));
+        }
 
-	/**
-	 * <p>
-	 * Left pad a String with spaces (' ').
-	 * </p>
-	 * 
-	 * <p>
-	 * The String is padded to the size of <code>size</code>.
-	 * </p>
-	 * 
-	 * <pre>
-	 * StringUtils.leftPad(null, *)   = null
-	 * StringUtils.leftPad("", 3)     = "   "
-	 * StringUtils.leftPad("bat", 3)  = "bat"
-	 * StringUtils.leftPad("bat", 5)  = "  bat"
-	 * StringUtils.leftPad("bat", 1)  = "bat"
-	 * StringUtils.leftPad("bat", -1) = "bat"
-	 * </pre>
-	 * 
-	 * @param str
-	 *            the String to pad out, may be null
-	 * @param size
-	 *            the size to pad to
-	 * @return left padded String or original String if no padding is necessary, <code>null</code> if null String input
-	 */
-	public static String leftPad(String str, int size) {
-		return leftPad(str, size, ' ');
-	}
+        if (pads == padLen) {
+            return str.concat(padStr);
+        } else if (pads < padLen) {
+            return str.concat(padStr.substring(0, pads));
+        } else {
+            char[] padding = new char[pads];
+            char[] padChars = padStr.toCharArray();
+            for (int i = 0; i < pads; i++) {
+                padding[i] = padChars[i % padLen];
+            }
+            return str.concat(new String(padding));
+        }
+    }
 
-	/**
-	 * <p>
-	 * Left pad a String with a specified character.
-	 * </p>
-	 * 
-	 * <p>
-	 * Pad to a size of <code>size</code>.
-	 * </p>
-	 * 
-	 * <pre>
-	 * StringUtils.leftPad(null, *, *)     = null
-	 * StringUtils.leftPad("", 3, 'z')     = "zzz"
-	 * StringUtils.leftPad("bat", 3, 'z')  = "bat"
-	 * StringUtils.leftPad("bat", 5, 'z')  = "zzbat"
-	 * StringUtils.leftPad("bat", 1, 'z')  = "bat"
-	 * StringUtils.leftPad("bat", -1, 'z') = "bat"
-	 * </pre>
-	 * 
-	 * @param str
-	 *            the String to pad out, may be null
-	 * @param size
-	 *            the size to pad to
-	 * @param padChar
-	 *            the character to pad with
-	 * @return left padded String or original String if no padding is necessary, <code>null</code> if null String input
-	 * @since 2.0
-	 */
-	public static String leftPad(String str, int size, char padChar) {
-		if (str == null) {
-			return null;
-		}
-		int pads = size - str.length();
-		if (pads <= 0) {
-			return str; // returns original String when possible
-		}
-		if (pads > PAD_LIMIT) {
-			return leftPad(str, size, String.valueOf(padChar));
-		}
-		return padding(pads, padChar).concat(str);
-	}
+    /**
+     * <p>
+     * Left pad a String with spaces (' ').
+     * </p>
+     * <p/>
+     * <p>
+     * The String is padded to the size of <code>size</code>.
+     * </p>
+     * <p/>
+     * <pre>
+     * StringUtils.leftPad(null, *)   = null
+     * StringUtils.leftPad("", 3)     = "   "
+     * StringUtils.leftPad("bat", 3)  = "bat"
+     * StringUtils.leftPad("bat", 5)  = "  bat"
+     * StringUtils.leftPad("bat", 1)  = "bat"
+     * StringUtils.leftPad("bat", -1) = "bat"
+     * </pre>
+     *
+     * @param str  the String to pad out, may be null
+     * @param size the size to pad to
+     * @return left padded String or original String if no padding is necessary, <code>null</code> if null String input
+     */
+    public static String leftPad(String str, int size) {
+        return leftPad(str, size, ' ');
+    }
 
-	/**
-	 * <p>
-	 * Left pad a String with a specified String.
-	 * </p>
-	 * 
-	 * <p>
-	 * Pad to a size of <code>size</code>.
-	 * </p>
-	 * 
-	 * <pre>
-	 * StringUtils.leftPad(null, *, *)      = null
-	 * StringUtils.leftPad("", 3, "z")      = "zzz"
-	 * StringUtils.leftPad("bat", 3, "yz")  = "bat"
-	 * StringUtils.leftPad("bat", 5, "yz")  = "yzbat"
-	 * StringUtils.leftPad("bat", 8, "yz")  = "yzyzybat"
-	 * StringUtils.leftPad("bat", 1, "yz")  = "bat"
-	 * StringUtils.leftPad("bat", -1, "yz") = "bat"
-	 * StringUtils.leftPad("bat", 5, null)  = "  bat"
-	 * StringUtils.leftPad("bat", 5, "")    = "  bat"
-	 * </pre>
-	 * 
-	 * @param str
-	 *            the String to pad out, may be null
-	 * @param size
-	 *            the size to pad to
-	 * @param padStr
-	 *            the String to pad with, null or empty treated as single space
-	 * @return left padded String or original String if no padding is necessary, <code>null</code> if null String input
-	 */
-	public static String leftPad(String str, int size, String padStr) {
-		if (str == null) {
-			return null;
-		}
-		if (isEmpty(padStr)) {
-			padStr = " ";
-		}
-		int padLen = padStr.length();
-		int strLen = str.length();
-		int pads = size - strLen;
-		if (pads <= 0) {
-			return str; // returns original String when possible
-		}
-		if (padLen == 1 && pads <= PAD_LIMIT) {
-			return leftPad(str, size, padStr.charAt(0));
-		}
+    /**
+     * <p>
+     * Left pad a String with a specified character.
+     * </p>
+     * <p/>
+     * <p>
+     * Pad to a size of <code>size</code>.
+     * </p>
+     * <p/>
+     * <pre>
+     * StringUtils.leftPad(null, *, *)     = null
+     * StringUtils.leftPad("", 3, 'z')     = "zzz"
+     * StringUtils.leftPad("bat", 3, 'z')  = "bat"
+     * StringUtils.leftPad("bat", 5, 'z')  = "zzbat"
+     * StringUtils.leftPad("bat", 1, 'z')  = "bat"
+     * StringUtils.leftPad("bat", -1, 'z') = "bat"
+     * </pre>
+     *
+     * @param str     the String to pad out, may be null
+     * @param size    the size to pad to
+     * @param padChar the character to pad with
+     * @return left padded String or original String if no padding is necessary, <code>null</code> if null String input
+     * @since 2.0
+     */
+    public static String leftPad(String str, int size, char padChar) {
+        if (str == null) {
+            return null;
+        }
+        int pads = size - str.length();
+        if (pads <= 0) {
+            return str; // returns original String when possible
+        }
+        if (pads > PAD_LIMIT) {
+            return leftPad(str, size, String.valueOf(padChar));
+        }
+        return padding(pads, padChar).concat(str);
+    }
 
-		if (pads == padLen) {
-			return padStr.concat(str);
-		}
-		else if (pads < padLen) {
-			return padStr.substring(0, pads).concat(str);
-		}
-		else {
-			char[] padding = new char[pads];
-			char[] padChars = padStr.toCharArray();
-			for (int i = 0; i < pads; i++) {
-				padding[i] = padChars[i % padLen];
-			}
-			return new String(padding).concat(str);
-		}
-	}
+    /**
+     * <p>
+     * Left pad a String with a specified String.
+     * </p>
+     * <p/>
+     * <p>
+     * Pad to a size of <code>size</code>.
+     * </p>
+     * <p/>
+     * <pre>
+     * StringUtils.leftPad(null, *, *)      = null
+     * StringUtils.leftPad("", 3, "z")      = "zzz"
+     * StringUtils.leftPad("bat", 3, "yz")  = "bat"
+     * StringUtils.leftPad("bat", 5, "yz")  = "yzbat"
+     * StringUtils.leftPad("bat", 8, "yz")  = "yzyzybat"
+     * StringUtils.leftPad("bat", 1, "yz")  = "bat"
+     * StringUtils.leftPad("bat", -1, "yz") = "bat"
+     * StringUtils.leftPad("bat", 5, null)  = "  bat"
+     * StringUtils.leftPad("bat", 5, "")    = "  bat"
+     * </pre>
+     *
+     * @param str    the String to pad out, may be null
+     * @param size   the size to pad to
+     * @param padStr the String to pad with, null or empty treated as single space
+     * @return left padded String or original String if no padding is necessary, <code>null</code> if null String input
+     */
+    public static String leftPad(String str, int size, String padStr) {
+        if (str == null) {
+            return null;
+        }
+        if (isEmpty(padStr)) {
+            padStr = " ";
+        }
+        int padLen = padStr.length();
+        int strLen = str.length();
+        int pads = size - strLen;
+        if (pads <= 0) {
+            return str; // returns original String when possible
+        }
+        if (padLen == 1 && pads <= PAD_LIMIT) {
+            return leftPad(str, size, padStr.charAt(0));
+        }
 
-	/**
-	 * Gets a String's length or <code>0</code> if the String is <code>null</code>.
-	 * 
-	 * @param str
-	 *            a String or <code>null</code>
-	 * @return String length or <code>0</code> if the String is <code>null</code>.
-	 * @since 2.4
-	 */
-	public static int length(String str) {
-		return str == null ? 0 : str.length();
-	}
+        if (pads == padLen) {
+            return padStr.concat(str);
+        } else if (pads < padLen) {
+            return padStr.substring(0, pads).concat(str);
+        } else {
+            char[] padding = new char[pads];
+            char[] padChars = padStr.toCharArray();
+            for (int i = 0; i < pads; i++) {
+                padding[i] = padChars[i % padLen];
+            }
+            return new String(padding).concat(str);
+        }
+    }
 
-	/**
-	 * <p>
-	 * Check if a String starts with a specified prefix.
-	 * </p>
-	 * 
-	 * <p>
-	 * <code>null</code>s are handled without exceptions. Two <code>null</code> references are considered to be equal.
-	 * The comparison is case sensitive.
-	 * </p>
-	 * 
-	 * <pre>
-	 * StringUtils.startsWith(null, null) = true
-	 * StringUtils.startsWith(null, "abc") = false
-	 * StringUtils.startsWith("abcdef", null) = false
-	 * StringUtils.startsWith("abcdef", "abc") = true
-	 * StringUtils.startsWith("ABCDEF", "abc") = false
-	 * </pre>
-	 * 
-	 * @see java.lang.String#startsWith(String)
-	 * @param str
-	 *            the String to check, may be null
-	 * @param prefix
-	 *            the prefix to find, may be null
-	 * @return <code>true</code> if the String starts with the prefix, case sensitive, or both <code>null</code>
-	 * @since 2.4
-	 */
-	public static boolean startsWith(String str, String prefix) {
-		return startsWith(str, prefix, false);
-	}
+    /**
+     * Gets a String's length or <code>0</code> if the String is <code>null</code>.
+     *
+     * @param str a String or <code>null</code>
+     * @return String length or <code>0</code> if the String is <code>null</code>.
+     * @since 2.4
+     */
+    public static int length(String str) {
+        return str == null ? 0 : str.length();
+    }
 
-	/**
-	 * <p>
-	 * Case insensitive check if a String starts with a specified prefix.
-	 * </p>
-	 * 
-	 * <p>
-	 * <code>null</code>s are handled without exceptions. Two <code>null</code> references are considered to be equal.
-	 * The comparison is case insensitive.
-	 * </p>
-	 * 
-	 * <pre>
-	 * StringUtils.startsWithIgnoreCase(null, null) = true
-	 * StringUtils.startsWithIgnoreCase(null, "abc") = false
-	 * StringUtils.startsWithIgnoreCase("abcdef", null) = false
-	 * StringUtils.startsWithIgnoreCase("abcdef", "abc") = true
-	 * StringUtils.startsWithIgnoreCase("ABCDEF", "abc") = true
-	 * </pre>
-	 * 
-	 * @see java.lang.String#startsWith(String)
-	 * @param str
-	 *            the String to check, may be null
-	 * @param prefix
-	 *            the prefix to find, may be null
-	 * @return <code>true</code> if the String starts with the prefix, case insensitive, or both <code>null</code>
-	 * @since 2.4
-	 */
-	public static boolean startsWithIgnoreCase(String str, String prefix) {
-		return startsWith(str, prefix, true);
-	}
+    /**
+     * <p>
+     * Check if a String starts with a specified prefix.
+     * </p>
+     * <p/>
+     * <p>
+     * <code>null</code>s are handled without exceptions. Two <code>null</code> references are considered to be equal.
+     * The comparison is case sensitive.
+     * </p>
+     * <p/>
+     * <pre>
+     * StringUtils.startsWith(null, null) = true
+     * StringUtils.startsWith(null, "abc") = false
+     * StringUtils.startsWith("abcdef", null) = false
+     * StringUtils.startsWith("abcdef", "abc") = true
+     * StringUtils.startsWith("ABCDEF", "abc") = false
+     * </pre>
+     *
+     * @param str    the String to check, may be null
+     * @param prefix the prefix to find, may be null
+     * @return <code>true</code> if the String starts with the prefix, case sensitive, or both <code>null</code>
+     * @see java.lang.String#startsWith(String)
+     * @since 2.4
+     */
+    public static boolean startsWith(String str, String prefix) {
+        return startsWith(str, prefix, false);
+    }
 
-	/**
-	 * <p>
-	 * Check if a String starts with a specified prefix (optionally case insensitive).
-	 * </p>
-	 * 
-	 * @see java.lang.String#startsWith(String)
-	 * @param str
-	 *            the String to check, may be null
-	 * @param prefix
-	 *            the prefix to find, may be null
-	 * @param ignoreCase
-	 *            inidicates whether the compare should ignore case (case insensitive) or not.
-	 * @return <code>true</code> if the String starts with the prefix or both <code>null</code>
-	 */
-	private static boolean startsWith(String str, String prefix, boolean ignoreCase) {
-		if (str == null || prefix == null) {
-			return (str == null && prefix == null);
-		}
-		if (prefix.length() > str.length()) {
-			return false;
-		}
-		return str.regionMatches(ignoreCase, 0, prefix, 0, prefix.length());
-	}
+    /**
+     * <p>
+     * Case insensitive check if a String starts with a specified prefix.
+     * </p>
+     * <p/>
+     * <p>
+     * <code>null</code>s are handled without exceptions. Two <code>null</code> references are considered to be equal.
+     * The comparison is case insensitive.
+     * </p>
+     * <p/>
+     * <pre>
+     * StringUtils.startsWithIgnoreCase(null, null) = true
+     * StringUtils.startsWithIgnoreCase(null, "abc") = false
+     * StringUtils.startsWithIgnoreCase("abcdef", null) = false
+     * StringUtils.startsWithIgnoreCase("abcdef", "abc") = true
+     * StringUtils.startsWithIgnoreCase("ABCDEF", "abc") = true
+     * </pre>
+     *
+     * @param str    the String to check, may be null
+     * @param prefix the prefix to find, may be null
+     * @return <code>true</code> if the String starts with the prefix, case insensitive, or both <code>null</code>
+     * @see java.lang.String#startsWith(String)
+     * @since 2.4
+     */
+    public static boolean startsWithIgnoreCase(String str, String prefix) {
+        return startsWith(str, prefix, true);
+    }
+
+    /**
+     * <p>
+     * Check if a String starts with a specified prefix (optionally case insensitive).
+     * </p>
+     *
+     * @param str        the String to check, may be null
+     * @param prefix     the prefix to find, may be null
+     * @param ignoreCase inidicates whether the compare should ignore case (case insensitive) or not.
+     * @return <code>true</code> if the String starts with the prefix or both <code>null</code>
+     * @see java.lang.String#startsWith(String)
+     */
+    private static boolean startsWith(String str, String prefix, boolean ignoreCase) {
+        if (str == null || prefix == null) {
+            return (str == null && prefix == null);
+        }
+        if (prefix.length() > str.length()) {
+            return false;
+        }
+        return str.regionMatches(ignoreCase, 0, prefix, 0, prefix.length());
+    }
 
 }
